Project 2/requirements.txt
deepgram-sdk
python-dotenv
sounddevice
numpy
elevenlabs
openai
